+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2024-02-03
 
+### Added
+- Display of Tokenexpiry on frontend when admin
+
 ### Changed
 - Added severity level to backend messages for improved clarity.
 - Reordered API structure by moving session API above the AuthMiddleware.